+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+
+- Retry dart pub upgrade on a git race in the pub cache
+
 ## 2.6.3 - 2026-09-23
 
 ## 2.6.2 - 2026-09-22